pyproject.toml
[tool.poetry]
name = "mastodon-autodelete"
version = "0.1.0"
description = "Automatically delete tagged posts after some timeperiod"
authors = ["'Jeong Arm' <'kjwonmail@gmail.com'>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.8"
lxml = "^4.9.1"
python-dateutil = "^2.8.2"
pytz = "^2022.1"
"Mastodonplus.py" = "^1.5.5.4"

[tool.poetry.dev-dependencies]
flake8 = "^4.0.1"
flake8-import-order = "^0.18.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
